chore(crypto): Move everything inside `crates/matrix-sdk-crypto/`.
